feat(Rewrite sniff for else on new line to be able to automatically fix itself)
Squashed commits:
fix(Space after if keyword does not work)

We should copy the whole logic of the Squiz.Sniffs.ControlStructure.ControlSignatureSniff because we still need the old rule inside this class. If we only keep the `else on new line` rule in the class, other rule will not be run because it get overwritten. (+1 squashed commits)

docs(Update code comment)

refactor(Update comment and remove unused variable) (+1 squashed commits)